test_task_2/requirements.txt
Flask==2.2.3
flask_sqlalchemy==3.0.3
psycopg2==2.9.6
